+++ R/c-grid.r
@@ -30,37 +30,17 @@
 #   ob
 # }
 
-#' Crop a ROMS layer
-#' 
-#' Crop a ROMS data layer from `romsdata` with a raster extent. 
-#' 
-#' The spatial crop is performed in the coordinate space of roms data. 
-#' @param x ROMS xy- coordinates, see `romscoords`
-#' @param ext `raster::extent` in the coordinate system of `x`
-#' @param ... ignored
-#'
-#' @export
-#' @examples
-#' ## notice that extent is in long-lat, but ice_local is in the grid
-#' ## space of ice_coords
-#' ice_local <- croproms(ice_coords, extent(100, 120, -75, -60))
-#' plot(ice_coords[[2]], col = grey(seq(0, 1, length  = 20)))
-#' plot(crop(ice_fake, ice_local), add = TRUE)
-croproms <- function(x, ext, ...) {
-  xy <- as.matrix(x)
-  incells <- which(xy[,1] >= xmin(ext) & xy[,1] <= xmax(ext) &
-                     xy[,2] >= ymin(ext) & xy[,2] <= ymax(ext))
-  x1 <- setValues(x[[1]], NA_real_)
-  x1[incells] <- 0
-  extent(trim(x1))
-}
-## https://www.myroms.org/wiki/easygrid
-# The rho-grid represents the grid centers (red dots)
-# The u-grid represents the grid East-West sides (blue triangles)
-# The v-grid represents the grid North-South sides (green triangles)
-# The psi-grid represents the grid corners (purple crosses)
 
+#'  https://www.myroms.org/wiki/easygrid
+#' The rho-grid represents the grid centers (red dots)
+#' The u-grid represents the grid East-West sides (blue triangles)
+#' The v-grid represents the grid North-South sides (green triangles)
+#' The psi-grid represents the grid corners (purple crosses)
+#' A useable Thredds example is here
+#' https://ecowatch.ncddc.noaa.gov/thredds/dodsC/northwest_fisheries/roms/nep_c.nc
 #' @importFrom raster crop 
+#' @name plot_cgrid
+#' @export
 plot_cgrid <- function(x, ex = extent(0, 15, 0, 20), 
                        include = c("u", "v", "rho", "psi"), cell = TRUE,  ...) {